requirements.txt
Flask
python-dotenv
google-api-python-client
